+++ resources/js/app.js
@@ -13,7 +13,8 @@ require('./bootstrap');
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-window.Vue = require('vue');
+import Vue from 'vue';
+window.Vue = Vue;
 
 // Notifications
 import Notifications from 'vue-notification';