+++ resources/lang/es/mail.php
@@ -12,15 +12,15 @@ return [
     'greetings' => 'Hola, :username',
     'want_reminded_of' => 'You wanted to be reminded of :reason',
     'for' => 'Para: :name',
-    'comment' => 'Comment: :comment',
+    'comment' => 'Comentario: :comment',
     'footer_contact_info' => 'Add, view, complete, and change information about this contact:',
     'footer_contact_info2' => 'See :name’s profile',
-    'footer_contact_info2_link' => 'Ver :name\'s profile: :url',
+    'footer_contact_info2_link' => 'Ver el perfil de :name: :url',
 
-    'notification_subject_line' => 'You have an upcoming event',
+    'notification_subject_line' => 'Tiene un evento próximo',
     'notification_description' => 'In :count days (on :date), the following event will happen:',
 
-    'stay_in_touch_subject_line' => 'Stay in touch with :name',
+    'stay_in_touch_subject_line' => 'Mantenga el contacto con :name',
     'stay_in_touch_subject_description' => 'You asked to be reminded to stay in touch with :name every :frequency day.|You asked to be reminded to stay in touch with :name every :frequency days.',
 
     'notifications_whoops' => 'Whoops!',
@@ -29,20 +29,20 @@ return [
     'notifications_footer' => 'If you’re having trouble clicking the ":actionText" button, copy and paste the URL below into your web browser: [:actionURL](:actionURL)',
     'notifications_rights' => 'Todos los derechos reservados',
 
-    'confirmation_email_title' => 'Monica – Email verification',
+    'confirmation_email_title' => 'Monica – Email de verificación',
     'confirmation_email_intro'=> 'Para validar tu email haz click en el botón de abajo',
     'confirmation_email_button' => 'Verificar la dirección de correo electrónico',
     'confirmation_email_bottom' => 'If you did not create an account, no further action is required.',
 
     'password_reset_title' => 'Monica – Reset Password Notification',
     'password_reset_intro' => 'You are receiving this email because we received a password reset request for your account.',
-    'password_reset_button' => 'Reset Password',
+    'password_reset_button' => 'Restablecer contraseña',
     'password_reset_expiration' => 'This password reset link will expire in :count minutes.',
     'password_reset_bottom' => 'If you did not request a password reset, no further action is required.',
 
-    'invitation_title' => 'Monica – You are invited by :name',
-    'invitation_intro' => 'You’ve been invited by :name (:email) to use Monica, a nice Personal Relationship Management tool.',
-    'invitation_link' => 'To accept the invitation, click on the link below:',
-    'invitation_button' => 'Accept invitation',
-    'invitation_expiration' => 'This link will expire in :count days.',
+    'invitation_title' => 'Monica – Has sido invitado por :name',
+    'invitation_intro' => 'Has sido invitado por :name (:email) a usar Monica, una buena herramienta de administración de relaciones personales.',
+    'invitation_link' => 'Para aceptar la invitación haz clic en el siguiente enlace:',
+    'invitation_button' => 'Aceptar invitación',
+    'invitation_expiration' => 'Este enlace caducará en :count días.',
 ];