+++ tests/Unit/Jobs/Reminder/NotifyUserAboutReminderTest.php
@@ -233,41 +233,4 @@ class NotifyUserAboutReminderTest extends TestCase
             'reminder_id' => $reminder->id,
         ]);
     }
-
-    /** @test */
-    public function it_creates_a_reminder_sent_object()
-    {
-        Notification::fake();
-        Carbon::setTestNow(Carbon::create(2017, 1, 1, 7, 0, 0));
-
-        $account = factory(Account::class)->create([
-            'default_time_reminder_is_sent' => '07:00',
-        ]);
-        $contact = factory(Contact::class)->create(['account_id' => $account->id]);
-        $user = factory(User::class)->create(['account_id' => $account->id]);
-        $reminder = factory(Reminder::class)->create([
-            'account_id' => $account->id,
-            'contact_id' => $contact->id,
-            'initial_date' => '2017-01-01',
-            'title' => 'fake text saying nothing',
-            'frequency_type' => 'year',
-            'frequency_number' => 1,
-        ]);
-        $reminderOutbox = factory(ReminderOutbox::class)->create([
-            'account_id' => $user->account_id,
-            'reminder_id' => $reminder->id,
-            'user_id' => $user->id,
-            'planned_date' => '2017-01-01',
-        ]);
-
-        NotifyUserAboutReminder::dispatch($reminderOutbox);
-
-        $this->assertDatabaseHas('reminder_sent', [
-            'account_id' => $user->account_id,
-            'reminder_id' => $reminder->id,
-            'user_id' => $user->id,
-            'planned_date' => '2017-01-01',
-            'nature' => 'reminder',
-        ]);
-    }
 }